Create QUE 3
This problem deals with matching mice to holes placed on a straight line. Since each mouse can move left or right in steps of one unit per minute, we need to place every mouse into a hole such that the maximum time taken by any mouse to reach its hole is as small as possible.