Create QUE 7
This problem deals with minimizing the difference between the tallest and shortest towers after modifying each tower’s height by either adding or subtracting a fixed value k. The challenge is to decide which operation to apply to each tower so that the final height difference is as small as possible.