Create QUE 10
This problem is about checking if a target string tar can be formed from a pattern string pat by simulating stack-like operations. For each character in pat, you can either append it to the result string s or delete the last character from s. The task is to determine whether it is possible to obtain tar after applying exactly one operation for each character of pat.