Create DAILY QUE 3
This problem focuses on finding the majority element in an array, i.e., the element that appears strictly more than n/2 times.

It can be efficiently solved using hashing to count frequencies or using Boyer-Moore Voting Algorithm for O(1) space. In this solution, we implement a custom hash table approach to count occurrences.