Create POTD QUE 21
The task is to generate binary representations of numbers starting from 1 up to a given number n. A binary number is written using only the digits 0 and 1, and each decimal number has a unique binary form. For example, if n = 5, the decimal numbers 1, 2, 3, 4, and 5 correspond to the binary numbers 1, 10, 11, 100, and 101. The goal of the problem is to produce this sequence of binary numbers in order.