Create POTD QUE 22
This problem is about simulating rotations on a deque. A deque supports efficient insertion and deletion from both ends, which makes it ideal for this task. In a right rotation, the element at the back is repeatedly moved to the front, while in a left rotation, the element at the front is repeatedly moved to the back. By performing these operations k times, we can achieve the desired circular rotation.