Create POTD QUE 23
We are given a binary array consisting only of 0’s and 1’s, along with an integer k. In one operation, we can choose a contiguous subarray of length k and flip all its bits (turn 0 into 1 and 1 into 0). The goal is to determine the minimum number of such operations required to make the entire array consist of only 1’s. If it is not possible to achieve this, we return -1. This problem essentially asks us to transform the array step by step, where flipping a window of size k is our only tool, and we want to find the least number of times to apply it.