Create DAILY QUE 4
Imagine you are counting votes in an election where each number in the array represents a candidate’s ID. The majority candidate is the one who gets more than half of the total votes. Your job is to figure out if such a candidate exists. If yes, return their ID; otherwise, return -1.