Create DAILY QUE 6
We are given an integer array and need to find the maximum possible sum of a subarray. A subarray is a continuous part of the array, and it must contain at least one element. The goal is to return the largest sum that can be formed from any such subarray.