Create TREE QUE 13
This solves the problem of finding the k-th ancestor of a given node in a binary tree. An ancestor of a node is any node on the path from the root to that node, excluding the node itself. The k-th ancestor is obtained by moving k steps upward in the tree. If no such ancestor exists, the function returns -1.

The approach builds an ancestor mapping for each node using a BFS traversal (storing each node’s parent). Then, starting from the target node, we repeatedly move to its parent k times to find the required ancestor.