Create DAILY QUE 14
You are given a fence with n posts and k colors. You need to find the total number of ways to paint the fence such that no more than two consecutive posts have the same color.
In other words, three consecutive posts can never be painted with the same color.