Create TREE QUE 29
In this problem, we are given the root of a binary tree and an integer targetSum. The task is to determine if there exists a root-to-leaf path in the tree such that the sum of all node values along that path equals targetSum. A leaf node is defined as a node that has no left or right child. The challenge is to carefully traverse the tree while keeping track of the path sum and checking if any path fulfills the condition.