Create POTD QUE 49
We are given a list of points on a 2D plane and an integer k. The goal is to find the k points that are closest to the origin (0,0) based on Euclidean distance. The distance for a point (x, y) from the origin is calculated as sqrt(x^2 + y^2). We can return the points in any order; the driver code will handle sorting if needed.