Create POTD QUE 63
In the bustling city of Worktown, each job brings profit — but time is money! Every job has a start time, an end time, and a profit if completed.
However, the city’s workers can only do one job at a time — they must finish one before starting another.

Your goal as the chief scheduler is to select a subset of jobs that don’t overlap in time and yield the maximum total profit.