Create DAILY QUE 47
In the world of Binaryville, every number is represented by a set of light switches — each bit is a switch that can be turned on (1) or off (0).
The ruler of Binaryville has a challenge: given a number n, turn all switches off (transform n into 0) using a special set of rules.
At each move, you can:
	1.	Flip the rightmost bit (bit 0).
	2.	Flip the i-th bit only if the (i−1)-th bit is 1 and all lower bits (from 0 to i−2) are 0.
The challenge? Do it in the minimum number of operations possible.
Your task is to determine how many moves it takes to transform n into 0.