Create POTD QUE 71
Imagine a long wooden stick marked from 0 to n. You need to cut the stick at certain marked positions. However, each time you make a cut, the cost is equal to the length of the current stick being cut. This means cutting earlier or later changes how large the remaining pieces are, and therefore influences future costs.  
Your goal is to decide the best order in which to perform all cuts so that the total cost of cutting the stick is minimized. This problem resembles a carpenter trying to minimize wasted effort by planning a cutting sequence carefully.