Create DAILY QUE 54
Imagine a long strip containing only zeros and ones. You are asked to examine every possible substring of this strip and determine how many of them show a clear dominance of ones.  
However, dominance is not determined by simple majority. Instead, a substring is considered to have dominant ones only if:
number of ones ≥ (number of zeros)²
This means that even a small number of zeros demands a comparatively large number of ones for the substring to be considered dominant.  
Your task is to count how many substrings satisfy this special dominance rule.