Create DAILY QUE 58
Imagine you are given a list of numbers and a starting number called original.  
Your task is to repeatedly search for original in the list:
• If it exists, you double the value of original.  
• If it does not exist, the process stops.

You keep doing this until the current value is no longer found in the array.  
The challenge is to determine the final value after all possible doublings.