Create DAILY QUE 60
Imagine you are given several intervals on the number line. Your task is to select a set of integers such that **each interval contains at least two numbers from your chosen set**.  
You want this set to be as small as possible.  
This is similar to covering each interval with two “pins” while using the fewest pins total.  
The challenge is to place these pins efficiently so that every interval is satisfied without wasting choices.