Create POTD QUE 78
Imagine a group of hospitals represented as nodes in a network. Some hospitals are directly connected through communication links, while others are isolated. You are allowed to take an existing link, disconnect it, and reconnect it between two **previously unconnected components** of the network.  
Your goal is to use as few such rewirings as possible to ensure that **all hospitals become part of one connected network**.
This problem is essentially about connecting multiple components of a graph into a single component using the links already available.